+++ pagerank.hpp
@@ -12,7 +12,7 @@ public:
      * Create a new PageRank object given a list of edges (represented as pairs
      * of node ids). Damping should usually be 0.85
      */
-    PageRank(const vector<pair<int, int>> edges, int numNodes, double damping);
+    PageRank(const vector<pair<size_t, size_t>> edges, int numNodes, double damping);
 
     /**
      * Get the result of the PageRank algorithm. Values are between 0 and 1.
